fix part 1

+++ src/test/java/SignUpTest.java
@@ -10,24 +10,41 @@ import pages.LoginPage;
 import pages.SignUpPage;
 import pages.StartPage;
 
-public class SignUpChromeTest {
+import io.restassured.response.ValidatableResponse;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.notNullValue;
+
+public class SignUpTest {
+
     private WebDriver driver;
+    private WebDriver yaDriver;
+    private String yandexBrowser = "C:\\Users\\dsladkov\\WebDriver\\bin\\yandexdriver.exe";
+    private String chromeBrowser = "C:\\Users\\dsladkov\\WebDriver\\bin\\chromedriver.exe";
 
 
     private By loginBtn;
     private String email;
+    private String password;
+    private String accessToken;
+    private UserClient userClient;
 
     @Before
     public void setUp() {
+        /**  Ревьюер предложил такой вариант: вынести путь до драйвера в строковую переменную */
 
+
+        System.setProperty("webdriver.chrome.driver", chromeBrowser);
         driver = new ChromeDriver();
         driver.get("https://stellarburgers.nomoreparties.site/");
+
+        userClient = new UserClient();
 
     }
 
     @Test
     @Description("Ошибка для некорректного пароля. Минимальный пароль — шесть символов")
-    public void SignUpIncorrectPassword(){
+    public void signUpIncorrectPassword() {
         StartPage startPage = new StartPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         SignUpPage signUpPage = new SignUpPage(driver);
@@ -36,10 +53,14 @@ public class SignUpChromeTest {
         loginPage.clickOnSignUpBtn();
         loginPage.waitToLoadSignUpPage();
         signUpPage.pasteNameSignUp(signUpPage.nameGenerator());
-        signUpPage.pasteEmailSignUp(signUpPage.emailGenerator());
+        loginPage.pasteEmailAuth(signUpPage.nameGenerator() + "@yandex.ru");
         loginPage.pastePasswordAuth("ljUoS");
         signUpPage.clickOnSignUpBtn();
         loginPage.checkErrorPassText();
+
+        UserCreds userCreds = new UserCreds(email, password);
+        ValidatableResponse loginResponse = userClient.login(userCreds);
+        accessToken = loginResponse.extract().path("accessToken");
 
     }
 
@@ -56,20 +77,23 @@ public class SignUpChromeTest {
         signUpPage.pasteNameSignUp(signUpPage.nameGenerator());
         email = signUpPage.emailGenerator();
         signUpPage.pasteEmailSignUp(email);
-        loginPage.pastePasswordAuth("ljUoSldk");
+        password = signUpPage.passGenerator();
+        signUpPage.pastePassSignUp(password);
         signUpPage.clickOnSignUpBtn();
 
-        UserCreds userCreds = new UserCreds(email,"ljUoSldk");
-        UserClient userClient = new UserClient();
+        UserCreds userCreds = new UserCreds(email, password);
         ValidatableResponse loginResponse = userClient.login(userCreds);
-        String accessToken = loginResponse.extract().path("accessToken");
-        userClient.delete(accessToken);
+        accessToken = loginResponse.extract().path("accessToken");
 
     }
 
     @After
-    public void tearDown(){
+    public void tearDown() {
         driver.close();
+        if (accessToken != null) {
+            userClient.delete(accessToken)
+                    .assertThat()
+                    .statusCode(202);
+        }
     }
-
 }
